Social Links in Profile
I added the ability to show a greyed social icon if no link is
provided, and a colored icon if one is provided. Also, the helper URL
text under the text field is updated with the user's information and
the icons are links to either the social platform, or the profile if
one is given.