Updated Devise Routes and Redirects
I updated the devise default routes for SEO purposes.

I also changed the default redirect after sign-in and edit profile from
the site_root to the user's profile for editing and the home page for
signing in.

Thanks to the following sites for information on how to do this:

- General information about Devise tweaks:

http://railscasts.com/episodes/210-customizing-devise?view=comments

- Changing URL pages for SEO purposes and easier user usage:

http://blog.railsupgrade.com/2014/05/how-to-override-devise-default-rout
es.html

- Setting the user_root page within the devise_for scope:

https://github.com/plataformatec/devise/wiki/How-To:-Customize-the-redir
ect-after-a-user-edits-their-profile

- Information on the after_update_path_for:

http://www.rubydoc.info/github/plataformatec/devise/Devise/Registrations
Controller#after_update_path_for-instance_method